instant/requirements.txt
boto3
keyring
bs4
pycrypto
python-dotenv
paramiko
pymysql
sshtunnel
click
requests